refactor(agents): extract AgentCapabilitiesDelegate from AgentEditorViewModel
Move the advanced-capability sections into AgentCapabilitiesDelegate: the
server-driven availability observers (code interpreter / web search / skills /
subagents, plus version-gated collaborative + handoffs/ACL) and the in-memory
section state (skill allowlist, subagent config, sequential chain, handoff
edges). The VM init starts them via observeAvailability(); the section handlers
are thin delegations.

No behavior change.
